modify tests
